@@ -1,6 +1,7 @@
 {
   "name": "svelte-vite-ssr",
   "version": "0.1.1",
+  "type": "module",
   "scripts": {
     "dev:csr": "vite --port 3000",
     "dev:ssr": "VITE_SVELTE_HYDRATABLE=1 PORT=3000 node ssr-dev-server",
@@ -10,35 +11,37 @@
     "serve:ssr": "PORT=3000 node dist/server"
   },
   "devDependencies": {
-    "@rollup/plugin-commonjs": "^19.0.0",
+    "@rollup/plugin-commonjs": "^21.0.1",
     "@rollup/plugin-json": "^4.1.0",
-    "@rollup/plugin-node-resolve": "^13.0.0",
-    "@rollup/plugin-typescript": "^8.2.1",
-    "@sveltejs/vite-plugin-svelte": "^1.0.0-next.12",
-    "@types/node": "^16.3.1",
+    "@rollup/plugin-node-resolve": "^13.1.3",
+    "@rollup/plugin-typescript": "^8.3.0",
+    "@sveltejs/vite-plugin-svelte": "^1.0.0-next.34",
+    "@types/node": "^17.0.10",
     "@types/serve-static": "^1.13.10",
-    "@typescript-eslint/eslint-plugin": "^4.28.3",
-    "@typescript-eslint/parser": "^4.28.3",
-    "@vitejs/plugin-legacy": "^1.4.4",
-    "eslint": "^7.30.0",
+    "@typescript-eslint/eslint-plugin": "^5.10.0",
+    "@typescript-eslint/parser": "^5.10.0",
+    "@vitejs/plugin-legacy": "^1.6.4",
+    "eslint": "^8.7.0",
     "eslint-config-enough": "^0.7.0",
     "eslint-plugin-ext": "^0.1.0",
-    "eslint-plugin-svelte3": "^3.2.0",
-    "postcss-load-config": "^3.1.0",
-    "postcss-preset-env": "^6.7.0",
-    "rollup": "^2.53.1",
+    "eslint-plugin-svelte3": "^3.4.0",
+    "postcss": "^8.4.5",
+    "postcss-load-config": "^3.1.1",
+    "postcss-preset-env": "^7.2.3",
+    "rollup": "^2.64.0",
     "rollup-plugin-import-meta-url-to-module": "^0.2.9",
     "rollup-plugin-string": "^3.0.0",
-    "svelte-preprocess": "^4.7.4",
+    "svelte-preprocess": "^4.10.2",
     "svelte-preprocess-css-hash": "^0.1.5",
-    "svelte-preprocess-html-asset": "^0.1.5",
+    "svelte-preprocess-html-asset": "^0.1.6",
     "svelte-preprocess-preload-link": "^0.1.0",
-    "typescript": "^4.3.5",
-    "vite": "^2.4.2"
+    "tslib": "^2.3.1",
+    "typescript": "^4.5.4",
+    "vite": "^2.7.13"
   },
   "dependencies": {
-    "serve-static": "^1.14.1",
-    "svelte": "^3.38.3",
+    "serve-static": "^1.14.2",
+    "svelte": "^3.46.2",
     "svelte-pilot": "^0.2.9"
   }
 }
